Commented out call to NoProject in script SayActiveCursor.
In SaySelectionPosition moved getting of field value after the key is sent to Audacity.  Previously Start/EndSelection was speaking the previous value.

Changed type of hPosition from Handle to Int.

Changed WrongFocus to test whether the window name of the focused control is not the track panel.

Updated messages and comments.  Also updated text in JSD, README, and What's new.

Removed commented-out code in IsMainWindow.

In script DeleteSelectedAudio moved Return to after SayNoProject.  Also added else so that key is sent if not in an audio track just in case it is used somewhere else.  This could also be done by moving TypeCurrentScriptKey at the start of the script.

In scripts Copy, CutToClipboard, and possibly others added additional conditions for doing the Audacity processing.

Copy/Cut audio work from the selection and tool bars.  The Copy and Cut scripts have been modified to handle this, but AnnounceMessage does not speak if !WrongFocus.  Did not change this because I don't know what affect this might have on other code.

Moved strings to announce Toolbar, Selection Bar, and Track Panel to audacity.jsm.

Made message strings in jsm file for many messages.  Did not do copy or cut.

Rewrote HandleCustomWindows.  Now announces when focus changes from the toolbars to the track panel or the selection bar.  Added IsToolBar, but it has not been tested and is not being used.

Scripts JawsHome andJawsEnd now do not speak the key name if they speak an Audacity message.  They do speak the key name if not in the track table or if Audacity announce messages are off.