AdjustJawsOptions, turn off toolbar announcement, silence Track Table, JSI file, changed function names of WrongFocus and IsMainWindow.
Added script AdjustJawsOptions and related code.  Added options for announce messages and announce toolbars.

Added script AdjustJawsVerbosity so that options will be available in versions prior to JAWS 10, not tested.  The Audacity options will appear at the top of the list but will not indicate that they are for Audacity.

It is now possible to turn off the announcement when moving to a new toolbar.

Added SayFocusedObject to not say "track table" when moving between tracks.

Changed file extension of settings file from .Ini to .jsi.

In SaySelectionPosition moved TypeCurrentScriptKey before the IsMainWindow test so that it will be sent to other dialogs.  Moved finding position inside AnnounceOn test since it is not needed until then.  Commented out sending key after a dialog opens because the sent key is what opened the dialog.

Changed Copy and CutToClipboard to not use AnnounceMessage.  This should allow them to speak in selection bar and toolbars.

Changed IsMainWindow to FocusInMainWindow.  Changed WrongFocus to FocusInTrackPanel, reversed the sense of its calls.

Added UserBufferIsActive to SayActiveCursor.

Made FocusInTrackPanel also call FocusInMainWindow so that condition does not need to be tested when it is used.

Rewrote script AnnounceOnOff to use UOAnnounceMessages.  It contains literal text.

HandleCustomWindows now calls HandleCustomWindows so that other functions such as in default.jsb can process the window.