audacity.jkm: Added ResetConfig script to laptop key settings.
audacity.jss:
Modified function SaySelection Position to make sure the selection toolbar is turned on.

Changed name of funtion WriteDefaultValue to AddDefaultConfig. Rewrote so that it can now be used to reset the script's settings to the default values.

Wrote script to resett script config to default, added its hotkey to the script hotkey help.

modified the function AutoStartEvent to write default settings if the config file does not exist.

audacity.jsd:
Changed function WriteDefaultValue to AddDefaultConfig.
Added script ResetConfig.