Silent uninstall, uninstalls existing version, bug fix in DisplayJawsListLeave.
Added /SD to some messageboxes in the uninstaller.  Will still prompt about changed files.

Added define UNINSTALLKEY containing the location of the uninstall info.

Added installdirregkey so that if the program is already installed, it will be found even if it is not installed in the default location.

Added code in .oninit to uninstall if the program is already installed.

Fixed a bug in DisplayJawsListLeave that was not providing the version string to CheckScriptExists.

If the user chooses not to overwrite scripts in a version, that version is deselected in the list view.

Added check for no versions selected, stays on the Select JAWS Versions page if that happens.