Silent uninstall, uninstalls existing version, bug fix in DisplayJawsListLeave, version from jss shown on welcome page.
Added define UNINSTALLKEY containing the location in the registry of the uninstall info.

Added installdirregkey so that if the program is already installed, it will be found even if it is not installed in the default location.

Added code in .oninit to uninstall if the program is already installed.

If the user chooses not to overwrite scripts in a version, that version is deselected in the list view.

Added check for no versions selected, stays on the Select JAWS Versions page if that happens.

In the JAWS Versions page now removes the unselected radio button from the tabbing order right after it is created.  The BS_Uautoradiobutton style causes the system to manage this when the buttons are changed, but both were initially in the tabbing order.

Adjusted positions of group box and radio buttons.

Added !define JAWSALLOWALLUSERS.  If not defined, the "Install for" group box is not included, SetShellVarContext is not executed, and the context value is not written to the ini file .

If JAWSALLOWALLUSERS is defined, the JAWS Versions page is shown even if there is only 1 JAWS version installed so that the user can choose whether to install to current user or all users.  If the JAWS Versions Page is displayed and there is only 1 version, that version is checked and initial focus is on the selected radio button.

Now remembers the previously selected radio button when returning to the page.

Added code to create and read the install INI file.

Added .jdf, .jfd, and .jff files to audacity installed items.

Added .jbf, .jbt, .jfd, and .jff to default script items.

Added define JAWSSrcDir which specifies the folder relative to the current folder on the system generating the installer of the folder containing the script files, ending with a backslash if nonempty.  Converted the Audacity files section and the default files function to use it.

Added code to get the script version from the script jss file.
Searches in ${JAWSSrcDir}${ScriptApp}.jss for the line:
Const CS_SCRIPT_VERSION = "<version>"
Case doesn't matter but spaces do.  VERSION is defined as the text between the double quotes.  If this line is not found, VERSION is set to "".

On the Welcome page now shows the title without a version if VERSION is not defined.

The sense of the overwrite scripts prompt was reversed.  This has been fixed.

Added /SD to some message boxes in the uninstaller.  Will still prompt about changed files.

Fixed a bug in DisplayJawsListLeave that was not providing the version string to CheckScriptExists.

Added define and macro to provide ${NSD_RemoveStyle} used to remove radio button from tabbing order.