Almost ready for separate header file.
Moved code to put all (or most) code that will go in the header file together.  Made code, pages, and sections into macros.

Changed the method of specifying what to do to install scripts in a JAWS version.  You now define the macro JAWSInstallScriptItems before including the header.  If not defined, a default version is used.

Made the definition of JAWSSrcDir overridable by specifying it before including the header.

Moved SetCompressor before the header section.  This is so that uninstlog.nsh could be used in code to install scripts in each version.

Moved SetOverwrite before the header.  This is so that the SetOverwriteDefault define is not in code that would be used to add to another installer.

Made the header section to be included only once, as if it were in its own file.