Installer source, README in full install, JAWSInstallFullItems.
Now provides a way for the user to specify files that should be installed in the install directory for full installs.

Issue: README and What's new are not made available with a just scripts install.

Widened list view by 2du.
Added access keys to radio buttons.

Added source files to the installer source section so that it now provides source for the full installer.

Changed README file define to use ${ScriptApp}.

Added macro JAWSInstallFullItems which generates a section to install items in the install folder like README, etc. for full installs.

Added a JAWSSec define and variable for the instdir files section and code in ComponentsPageLeave to select it when type is not Just Scripts.