README in install-dependent location.
jfw.nsh:
Added a JAWSREADME variable that is referenced by the default MUI_FINISHPAGE_SHOWREADME define to tell the Finish page where the README file is.  This allows it to be installed in different places based on the install type.

Added a JAWSNoReadme define which suppresses the default definition of MUI_FINISHPAGE_SHOWREADME.

installer.nsi:
Added code to install the text files to the script folder for Just Scripts installs.  Uses the $JAWSREADME variable to tell the Finish page where the README file is.  For a full install, it also sets this variable.

Now gets the README and Whats new from the folder containing the scripts.

Added commented out define of JAWSNoReadme which suppresses the default definition of MUI_FINISHPAGE_SHOWREADME.