License page and legal copyright.
JFW.nsh:
Added license page, included if JAWSLicenseFile is defined.  JAWSLicenseFile is the name in $JAWSSrcDir of the license file.
Now Shows the contents of ${LegalCopyright} on the Welcome page if defined.

installer.nsi:
Added the JAWSLicenseFile and LegalCopyright defines.  The license file is defined as COPYING.TXT in the ${JAWSSrcDir} folder.