Changed IsStopped to use GetAudacityState.  This is because on Cuong's machine the Stop button does not return CTRL_DISABLED.
This version does not identify object states on Cuong's machine.  It contains code that fixes this, but it executes far too slowly.  Intend to try to do it by checking the name of the button graphic.

Added Cuong's SelectAll script, but changed FocusInTrackPanel to FocusInMainWindow because it works in selection bar and toolbars as well.

Removed duplicate line for AdjustJawsOptions in hotkey help.

Added and commented out code in GetObjectState to get object state with invisible cursor.  This works on Cuong's machine, but it is very slow, so commented out.

Merged in Cuong's change to make SelectFromStartOfLine and SelectToEndOfLine scripts now warn when no project is open.