Audacity state by toolbar graphics.
In GetAudacityState added tests of the toolbar button graphic name.  If the first call to GetObjectState fails, uses the graphics to identify the state of the remaining controls.  Added CS_IMG constants for the pressed state of the Play, Pause, and Record buttons.

Added FocusInTrackPanel test to condition for speaking position in FinishMarkerLeft and StartMarkerRight.

Added no project message in JAWSHome and JAWSEnd.