Merged missed changes from br_69fix.
Deleted duplicate message in .jsm file reintroduced in merge after it was deleted in rev 67.
Scripts SaySelectionStart, SaySelectionEnd, SayAudacityState, and DeleteSelectedAudio now pass keys to app when in a dialog.

In SayAudacityState reversed the order of ActiveDialog and no transport toolbar tests.

In SaySelectionPosition reversed the order of DialogActive and NoProject tests.

Removed frame (.jfd and .jff) files since they are no longer used.