IsTrackSelected works when focus not in track panel, SayNotStopped.
Made function SayNotStopped which speaks the msgNotStopped_l/s message, honoring the setting of the Announce Audacity Messages option.

Fixed the following issue: If DEL is pressed when focus is not in the track panel the script says no track is selected and says to press enter to select it.  This is because IsTrackSelected looked at the object at the PC cursor.  This means that it would not allow you to delete if a track is selected and the focused track is not.

Added code in IsTrackSelected that doesn't require that the track panel have focus.  IsTrackSelected now returns True if at least one track is selected.  Changed message to not mention pressing ENTER since we don't know if the focus is in the track panel or on the desired track.

Reorganized and modified What's new file.  Changed to reflect that the ENTER key pauses while playing/recording feature is enabled by default.

Removed [laptop keys] section and debugging keys from .jkm.