Changed the name of SwichPane to SwitchChainsList in jkm, jsd, jsm.
In ProcessVST does not turn speech on if it was previously off.  Added pause after setting focus so that button is not spoken.

In SwitchChainsList changed so that focus is placed on chains list if focus is on any control in the dialog except the chains list.

Removed "to close" line from hotkey help.  Added call to AddHotkeyLinks (defined in default.jss) to a link to display default JAWS hotkeys.  There is also commented out code to add the link and clos message in case AddHotkeyLinks doesn't work.
The link for displaying default JAWS hotkeys in Hotkey Help works from the links list (JAWSKey+F7) in the virtual viewer.
In JAWS 10 when the link is activated from the text in the virtual viewer, it redisplays the Audacity hot keys list.  This is the same behavior as seen in Outlook for Office 2003.