Installer: JAWSALLOWALLUSERS fixes
In jfw.nsh:
Added comment on declaration of JAWSSHELLCONTEXT indicating where its default value is set.
Changed PageInstConfirmPre so that if JAWSALLOWALLUSERS is undefined, it indicates that scripts will be installed for the current user.

Added documentation for JAWSALLOWALLUSERS and a warning if it is not defined.  Removed ifdef in PageInstConfirmPre around adding current or all users.  This allows the including file to redefine JAWSSHELLCONTEXT after including this file.  It also requires that JAWSSHELLCONTEXT is set even if JAWSALLOWALLUSERS is undefined.
Commented out debug message displayed when JAWSALLOWALLUSERS is not defined and there is only one JAWS version installed.