Link to JAWS guide web page.
Added code to make the link to the Audacity JAWS guide an active link.  Split the msgAudacityHotKeyHelp into two messages-- before and after the link.  The code is designed so that it can be modified to read the URL from the JSI file.

Changed script Enter  to perform the default enter script in the user buffer.  This causes links to be run by the ENTER key.

Changed so that Audacity hotkey help does not display a link to default JAWS hotkey help.

The link in Audacity hotkey help now displays the default JAWS hotkey help instead of itself over again.  Changed script HotKeyHelp to AudacityHotKeyHelp and WindowKeysHelp to AudacityKeysHelp.  Added definitions for them in jkm file.