Improved speaking in effect dialogs.
Now speaks the value associated with a slider in an effect dialog as the slider is changed.  (Changed HandleCustomWindows and added SayNonhighlightedText.)
Also explicitly gets the window name of the edit control as the slider name, which cuts out chatter in some dialogs.

This has been tested in some Nyquist effect dialogs, but hasn't been extensively tested in all dialogs.