Attempted fix for finding scompile on win64.
In .oninit we now check to see if $programfiles\Freedom Scientific\JAWS exists.  If it does we set new variable $JAWSPROGDIR to it.  Otherwise we check for it in $programfiles64.  If neither of those are found we write a message saying we may not be able to compile the scripts.  (The only way we could would be if we can get the location from the registry.)
 Also reactivated the message when the compile failed.  This might have been deactivated because it will happen for each JAWS version.