Added Insert+ keys.
In jss, jkm, and jsd changed AudacityHotkeyHelp to AudacityScriptKeyHelp.  This name more clearly (to me at least) indicates what the script does, and it removes a second (bogus) entry in Keyboard Manager.

In audacity.jkm added an Insert entry for each JAWSKey entry.  The intent is to make the Insert key still work in Laptop mode when the JAWSKey is specified as CapsLock.