Adds Quick Settings support and ability to change userr guide URL.
Commented out keystroke Jawskey+V for testing the quick settings script.

Wrote xml files for the QuickSetting (Similar to script adjust Jaws options). This is for use on Jaws 13 or later.

Wrote function LoadNonJCF Options to get info from JCF file, and write to Audacity>JSI file. This is for script QuickSettings

Modified Hotkey Help message. Now the hotkey Insert+v displays in all Jaws versions, although it was removed from the JKM file.

Modified function HandleCustomWindows. Jaws now reads the checkbox and button in the quick settings dialog when navigated to these by the tab key.

Make the settings section in the JSI file as a constant so it is easier to change if needed, and modified all related scripts.

Added script AddAudacityJawsGuide for changing Jaws guide link.
Script returns if new and current URL are the same or if the user does not enter anything in the box.
Updated message and documentation.

Updated messages.