QuickSettings fixes.
Changed AnnounceOn to GetQuickSetting.  It now uses the JSI file for JAWS prior to 13.0 and GetNonJCFOption for 13.0 and later.

Decided that for JAWS 13 or later we use QuickSettings and NonJCFOptions, and we don't attempt to transfer settings from a JSI file.  For prior to JAWS 13 we continue to use JSI file.

In AutoStartEvent made loading of JAWS guide link only execute for pre JAWS 13, and added initialization for it in LoadNonJCFOptions.

Added new globals gsIniFile and gsIniSection.  For pre JAWS 13 gsIniFile is the old CS_Ini_File, for 13 and later it is "Audacity.jcf".  gsIniSection is "Settings" and "NonJCFOptions", respectively.  Removed constant Settings, which was replaced by gsIniSection.  Most existing code that only executes pre JAWS 13 still uses CS_Ini_File.  These changes should allow AddAudacityJAWSGuide to work.

In script AddAudacityJawsGuide now initializes the URL to the default if not in file.

In Audacity.qsm corrected case of ID for the root node.  The root node text and help text are now correctly displayed.