Silence Preview.
Added code to silence speech during previewing.  The flags associated with this (gfSilence and gfPreviewing) are cleared by AutoStartEvent every time it runs, so if they aren't turned off when previewing is over switching focus away from Audacity and back will clear them.  However, have not been able to get this implementation stuck.

Implemented a flag to suppress the next tutor message.  This is currently not used.

Changed 1 to symbolic constants in code that writes default settings entries and the default value of code that reads settings.