Added Open/CloseListBox.
They pass the key through when focus is in the track panel.  This is because the key wasn't being sent through which causes problems if it is bound to a command such as MoveTrackUp/Down.

Added code for the Test script code from the file used to test the go to track functionality.