Added foreign language structure to repo and Spanish translation.
Added lang and lang/esn folders, and Spanish translations for the 9/14/13 release.

In audacity.jsm moved the UO_*Hlp messages following the other UO constants as in the English file.

Added the current qsm file with text copied from the Spanish translations of the equivalent messages from the Spanish JSM.  Newer messages are included untranslated.