Install scripts with multiple languages.
The JAWS versions list view now has an entry for each installed version/language.
Unsupported version/language pairs  are marked with a trailing * in the list view, and the scripts for the default language are used if they are selected.

One compile seems to work, with multiple compiles it seems like only 1 works, although the files are extracted for all of them.  If you run separate installs the compiles workk.

JFW.nsh:
Added define ScriptVerLangSep to separate version and language directory.  Removed enu from ScriptDir.

Added qs and qsm files to default JAWSInstallScriptItems macro.

Converted GetJawsScriptDir to work with a version/language pair or just a version.

Converted CompileSingle to work with version/lang pair.

Converted keywords in JAWSSaveInstallInfo and JAWSRestoreInstallInfo to JAWSVersionLangs and JAWSVersionLangsCount.

Macro JAWSInstallItems now receives version in $0 and lang in $1.

Wrote and tested ${StrContainsTok}, used to see if an installed language is in the supported languages list.

Widened listview to accomodate language.

Commented out can't find compiler message in InstallJawsVersion which duplicates one in CompileSingle.

Further documented list view code.

Changed DEFAULT_GUI_FONT to GUI_DEFAULT_FONT, which removed 1 warning.

Added DumpLog function, called in leave function of instfiles page.  Writes the log window to installer.log in the $InstDir.

installer.nsi:
Added constructs for languages, and Spanish language message files.
Added QuickSettings files.

With this version the list view did not speak properly in JAWS 16.0.4463. It later began to work, no known changes.  Items spoke when SayLine was pressed but not for prior/next line.  The checked state was not spoken for either of these, only when SPACE was pressed to change state.  It did work in NVDA 2015.3.

Issue:  I observe that after running the installer from File Explorer, the focus has moved a couple of tabs to the right-- i.e. to View Mode or something.