2.1.2RC1 support.
Contains fix for Silence Preview for 2.1.2RC1.

Previewing for 2.1.2RC1 is detected in HandleCustomRealWindows by watching for the strings "Previewing" and "Preparing preview", which is in the first child of the Previewing progress dialog.  The Stop button is detected by control ID and window name.  Control ID can't be used by itself because it is the same as OK in some dialogs.  Control ID is probably redundant.

HandleCustomAppWindows now just returns True.  This suppresses chatter related to previewing.  We shouldn't be switching apps anyway, and this seems to work okay, at least in 2.1.2RC1.

Added NewTextEventShouldBeSilent to suppress speech in Silence Preview.  Didn't solve the problem it was added for, but a good idea anyway.  The new method of handling Silence Preview doesn't work in Audacity 2.1.1 (but the 2.1.1 code is still there so the script still works on 2.1.1).

Added gfAudacityAutostarted to deal with switching from another app when focus is in a dialog, which triggered the Silence Preview logic.  It is cleared at the start of HandleCustomWindows.  It was needed when testing for previewing was done in HandleCustomAppWindows, and is probably no longer be needed since testing for previewing is now done in HandleCustomRealWindows.

Made constants for previewing strings and stop button.

Tested several times in Noise Reduction and Amplify without problems.

Have observed the Stop button speaking when previewing starts.  This seems to be spoken in the code that speaks the focused window, which is shut off when the real window for the preview dialog is detected.  That suggests that the Stop button is gaining focus before the real window is detected, which shouldn't happen.