Changed readme to markdown.
Changed source files audacity_readme.txt and audacity_readme_vi.txt to readme.md and readme_vi.txt.  Changed build.cmd to copy these to readme.txt and readme_vi.txt in the build folder.  Changed installer.nsi to process these.

This allows the readme to be displayed by GitHub.  (Should the readme displayed by GitHub be the product readme or one oriented to users of the GitHub repo, which might link to the product readme?)