Fixed inclusion of license page.
The license page was previously not included.  This is because !ifdef does not recognize LangStrings.

Updated submodule jfw_nsh.  It now requires that LicenseLangString be used for the license file, and that there be a define for JAWSLicenseFile as well to cause it to be included.
Language files now use LicenseLangString and include the path to the license file.
installer.nsi: The langstring name is not used to copy the license file.  It is now hard-coded in JAWSInstallFullItems.