Changes for Audacity 2.1.3-- toolbars, record/play meters.
Changed Next/PriorDocumentWindow to support toolbars in Audacity 2.1.3 (audacity-win-re503358-2.1.3-alpha-10-nov-16).

Changed SayPlayback/RecordMeter to support v2.1.3 meters.

Code is still in SayRecordingMeter to say playback meter if playing.  Was this intentional?  If so, play meter code hasn't been converted properly.

Combined meter is undocked.  Was able to get it to work by editing audacity.cfg (uisng audacity-win-r396a6f0-2.1.3-alpha-26-nov-16).

When PreviousDocumentWindow  reaches the device toolbar it says "device panel" instead of Playback Device because there is a visible enabled panel there.  Don't know what it's for.  It is not in the tabbing order.