Added algorithm for tempo that uses average of last 8 times commented out.
This algorithm is based on Robert Hänggi's algorithm in his NVDA add-on.  He sets the initial tempo average to 200BPM.  This has not been factored into this algorithm.  They are from branch tempo_array but the 2 is removed from the script names. This version of the scripts are commented out.

Broke lines in script key help for tempo so that they aren't truncated in user buffer.

In audacity.jsm changed script version in initial comment to 2.2.0.

Changed program version to 2.2.0-Alpha-2017-08-15.