Updated CONTRIBUTING.md.
Made separate lists for adding a language to the script (including adding it to the installer) and adding a new language to the installer interface.

The information is changed, but there are some formatting problems, at least with md2html.  In the section "Adding a new language to the installer" md2html connects the list with the explanations of the dates (which is actually supposed to be a ntested list) with the list following it (which should be the continuation of the main  list.  This happens even though there are 3 blank lines between the list items.

The comment block should actually be part of the above list item but I can't get md2html to do that.  md2html requires that ``` code fence markers be in column 1, and indenting the code block doesn't include <pre> around the block.  What GitHub will do with this I can't tell without pushing the file.