Added code to prreview with cursor motion.
Added script ToggleMotionPreview assigned to key JAWSKey+p.  It toggles motion previewing on and off and Say Position between All and All but Arrows.  It does not change the Quick Setting.  It reverts to the Quick Settings values when Quick Settings is opened or when focus switches away from Audacity and back.  The message constant for the pre-JAWS 13 Adjust Options is there but the rest of the support is not.  Have added it to JSD and Script Key Help.

Updated What's New.

Changed program version to 2.2.0-alpha-2017-08-27.