Changes to audio previewing keys.
Changed sayPrior/NextWord to speak after the start and before the end (inside the time range) and SelectPrior/NextWord to speak before the start and after the end (outside the selection).  These now also work in the selection bar.  Updated JSD and script key help.

Changed so that commands that move the start cursor speak after it and those that move the end of the selection speak before the end.  This means that what is spoken is always what will be played by SPACE.

JAWSKey+Left/RightArrow, etc. now also preview in the selection bar.

Updated What's New.

Changed program version to 2.2.0-alpha-2017-08-31.