German translation for 2.2.0 first try.
Updated source file variables in build.cmd.

The German audacity.jsm says it is based on the 8/24 version but this version contains the 9/1 version.  lang/deu/audacity.jsm appears to be based on commit 6dbf582.

Added the missing constants with English messages so that the script will at least compile.

Changed the comment at the start of deu/audacity.jsm so that it says "translation dated" so that the first line contains the date of the English file so that it can be copied verbatim.

Updated What's New.