Updated Spanish translation:
Note: The English files in this revision are not up-to-date, should have merged next before committing these.  jfw_nsh shows as modified, did not sync submodules, probably not important as long as you don't generate an installer from this revision.

Changes by Fernando Gregoire:
I've just updated the files I was missing compared to the latest changes. I've also fixed a mistake in the Spanish JKM related to the ToggleMotionPreview script.

Regarding the Read Me, I updated the date and things related to the Vietnamese Read Me.