Added IDs for selection bar for Audacity 2.3.0-rc2.
Added globals for these IDs, modified GetPositionFieldID and SaySelectionStart/End.  We now use GetPositionFieldID in SaySelectionStart/End.

Added string WN_MANAGE_MACROS and messages for macros.  In script SwitchChainsList added code to handle v2.3.0.

Equalization dialog speaks DB when tabbing between between bands, but not when arrowing a slider.

Silence Preview doesn't work in 2.3.0, or it silences but the silence flag is not cleared.
.

Version 2.2.2-beta-2018-10-25.