Added code to handle speaking position after cursor motion.
Added functions FormatPositionField and SchedulePositionField.  Also added global ghSayPositionField and gfPositionFieldUpdated and code in SayNonHighlightedText and added ScreenStabilizedEvent.  In script SayJump SchedulePositionField appears before TypeCurrentScriptKey.  This makes it work all the time, but we need to deal with the checks that follow.  One way to do this is to clear ghSayPositionField before error returns, which we currently do.  What do not work are SelectionStart and SelectionEnd which use SaySelectionPosition.