Added code in SaySelectionPosition to use SchedulePositionField.
This works for SelectionStart but does not seem to work well for SelectionEnd.  The SchedulePositionField method of speaking changed selection bar values does not work if GetPositionField does not contain formatted results (because punctuation was not returned).  It also helps to maximize the window.

Updated copyright in audacity.jss.

Rev 2.2.2-beta-2020-01-05.