Selection expansion/contraction keys now speak every time.
In Start/FinishMarkerRight/Left moved SchedulePositionField before call to MarkerMovement.  These work, but the position takes a long time to speak, and pressing another one before the position speaks does not interrupt speaking the prior position.  This may be a problem.  Tried to fix by clearing gfPositionFieldUpdated in SchedulePositionField.  This is a good idea anyway, but it did not affect the speaking of the keys.

Rev 2.2.2-beta-2020-05-29B.