Was for debugging, but okay to keep.
This was a debug version for finding out why the selection bar could not be found.  It turned out that I was running Audacity V3.2.3,.  The debugging has been turned off, but there were a few updates of the SayStrings that are worth saving.

Also, the selection bar can't be found in 3.2.3, its ID is 9.  Why??

Using JAWS 2024.

Rev: 2.2.2-beta-2024-03-15.